post-processing-microservice/requirements.txt
pytest
kafka-python